Hide delete in view barangay when not admin

+++ resources/views/senior_citizen/view_barangay.blade.php
@@ -29,8 +29,13 @@
                           <td>{{$list->middlename}}</td>
                           <td>{{$list->sex}}</td>
                           <td>{{$list->birthdate}}</td>
-                          <td><a href="/edit_citizen/{{$list->id}}">Edit</a> 
-                            <a href="/delete_citizen/{{$list->id}}" id="delete_confirmation">Delete</button></a>
+                          <td>
+                            <a href="/edit_citizen/{{$list->id}}">Edit</a> 
+                            <a href="/view_citizen/{{$list->id}}">View</a> 
+                            @if(auth()->user()->position == 'Admin')
+                                <a href="/delete_citizen/{{$list->id}}" id="delete_confirmation">Delete</button></a>
+                            @endif
+                          </td>
                       </tr>
                     @endforeach
                 </tbody>
